+++ src/db/pgSQL.ts
@@ -1,7 +1,15 @@
-import {FormattedSqlValueExpression, rawSql} from '../query/SQLExpression';
+import {
+  BaseSqlExpression,
+  FormattedSqlValueExpression,
+  QueryContext,
+  rawSql,
+  value
+} from '../query/SQLExpression';
 import {IDBDialect, setDbDialect} from './index';
 import {ISelectStatement} from '../query/statements';
-import {IFieldReference, ISQLExpression} from '../dbTypes';
+import {IFieldReference, IQueryContext, ISQLExpression} from '../dbTypes';
+import {DataValue, MAX_SINGLE_LINE_STATEMENT_LENGTH} from '../query/types';
+import indentString from 'indent-string';
 
 const encryptFormatFn = (fldText = '') =>
   `encode(pgp_sym_encrypt(${fldText}, $[encryptionKey]), 'hex')`;
@@ -17,6 +25,40 @@ const hashPwFormatFn = (fldText = ''): string =>
 
 const hashPwValFormatFn = (fldText = '', fldRefText = '') =>
   `crypt(${fldText}, ${fldRefText})`;
+
+interface ICoalesceSqlExpression {
+  expressions: ISQLExpression[];
+}
+
+class CoalesceSqlExpression extends BaseSqlExpression
+  implements ICoalesceSqlExpression {
+  public expressions: ISQLExpression[];
+
+  constructor(...expressions: Array<DataValue | ISQLExpression>) {
+    super();
+    this.expressions = expressions.map(expression => value(expression));
+  }
+
+  isSimpleValue = () => {
+    return this.expressions.every(expression => expression.isSimpleValue());
+  };
+
+  toSql = (qryContext: IQueryContext = new QueryContext()) => {
+    if (this.expressions.length < 2) return '';
+    const values = this.expressions.map(expression =>
+      expression.toSql(qryContext)
+    );
+    if (
+      values.reduce((totChars, valueLine) => totChars + valueLine.length, 0) +
+        10 +
+        2 * (values.length - 1) <
+      MAX_SINGLE_LINE_STATEMENT_LENGTH
+    ) {
+      return `coalesce(${values.join(', ')})`;
+    }
+    return `coalesce(\n${indentString(values.join(',\n'), 2)}\n)`;
+  };
+}
 
 export class PgDialect implements IDBDialect {
   public encryptField = (value: ISQLExpression): ISQLExpression =>
@@ -55,6 +97,11 @@ export class PgDialect implements IDBDialect {
     }
     return selectStatement;
   };
+
+  nullValue = (
+    val1: ISQLExpression | DataValue,
+    val2: ISQLExpression | DataValue
+  ): ISQLExpression => new CoalesceSqlExpression(val1, val2);
 }
 
 export const usePg = () => {